+++ changelog.md
@@ -1,4 +1,5 @@
 0.9.29 (in development)
+- Fixed: d1_trainstation_05: players able to escape the teleport sequence.
 
 0.9.28
 - Improved: Better prediction handling for the Gravity Gun, should be more responsive with higher ping now.